mm: kasan: unified support for SLUB and SLAB allocators
With this patch kasan can be compiled with both SLAB and SLUB allocators,
using minimal dependencies on allocator internal structures and minimum
allocator-dependent code.

Dependency from SLUB_DEBUG is also removed. The metadata storage is made
more efficient, so the redzones aren't as large for small objects. The
redzone size is calculated based on the object size.

Stack storage isn't implemented here and is planned for a future patch

The gfp flags are added to kasan hooks for future patches as well.

Signed-off-by: Dmitry Chernenkov <dmitryc@google.com>